pyproject.toml
[tool.poetry]
name = "cofrinho"
version = "0.1.0"
description = ""
authors = ["dunossauro <mendesxeduardo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"
ipdb = "^0.13.4"
pyhamcrest = "^2.0.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
